requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
colorama==0.4.6
connexion==3.2.0
coverage==7.8.0
cryptography==44.0.2
distlib==0.3.9
dj-database-url==2.3.0
Django==5.2
django-bootstrap5==25.1
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-environ==0.12.0
django-heroku==0.3.1
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.18.0
flake8==7.2.0
Flask==3.1.0
Flask-Testing==0.8.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
lxml==5.3.2
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
oauthlib==3.2.2
openapi-client==1.1.7
packaging==24.2
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
py==1.11.0
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
PyJWT==2.9.0
pyproject-api==1.9.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
pytest-mock==3.14.0
pytest-randomly==3.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
swagger_ui_bundle==1.1.0
tomli==2.2.1
tox==4.25.0
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.30.0
Werkzeug==3.1.3
whitenoise==6.9.0
